2.1.15-1 : Fix 2054660 : hang with multiple Broadcast and multiple rings
This bug was reported with PyTorch. When we do not enqueue work on all
rings with saveProxy, startProxy might think later that there was work
posted because the FIFO is full and we are looking at an old request.
startProxy is therefore incrementing the tail and duplicating a
proxy operation, that extra operation causing a hang.

Also, changing proxy head/tail to uint64_t to avoid problem if we ever
do more than 2GB proxy operations (and be consistent with 2.2.x).

+++ src/include/transport.h
@@ -63,8 +63,8 @@ struct transportProxyInfo {
   threadFunc_t func;
   volatile int proxyReady;
   struct ncclProxyArgs argsFifo[TRANSPORT_PROXY_FIFO_SIZE];
-  volatile int argsFifoHead;
-  volatile int argsFifoTail;
+  volatile uint64_t argsFifoHead;
+  volatile uint64_t argsFifoTail;
   pthread_cond_t cond;
   pthread_mutex_t mutex;
 };
